Added helper class PandoraStation and helper methods
This helper class is used to store the components of the pandora
stations and any additional information I would want to display on
the GUI side of things. A helper method was created that iterates
the linked list of stations and puts them into a vector of
pandora station objects.