Ok so we are ditching libcurl because... well its a pain
Now we are going to directly access the music using QUrl and letting
the phonon backend deal with it. This makes the program KDE dependent,
but by choosing QT I was already going for that, so it doesn't matter.

Summary:
The program will log into the system and download the first song from
the playlist and play it until the end. The program then just sits.