Fix the anoying and unpredictable memory errors, elapsed time
When it comes to string manipulation, I think I should leave that
to the experts and just use the QString library. Using this has
eliminated all my problems so far. It seems like it wasn't all
libcurls fault, but now it's too late to got back. In a good way this
problem forced me to use a more elegant solution, Phonon. On the bad
side I wasted about 5 hours or more trying to figure this out.

Changes:

The app runs, plays the first song and displays the elapsed time.