Song Information Display Added, seek bar added
The song information is now displayed bellow the album art. At the very
bottom of the screen is a seekbar for the media playing. Doesn't
seem to work, kind of crashes the program. Worse case, it should be
just used as a display with no real use, other than looks.